final/django-api/requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.39.14
botocore==1.39.14
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
iyzipay==1.0.45
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
